Refactor FindPreprintPDF
